requirements.txt
svg.path
numpy
turtle
pyserial
structlog
